visualization/backend/requirements.txt
pydantic==1.10.13
fastapi==0.104.0
uvicorn==0.23.2
tortoise-orm==0.20.0
aiomysql==0.2.0